:recycle: refactor: calculate 컴포넌트 분리
